blog_project/src/apps/categories/serializers.py
from rest_framework import serializers

from src.apps.categories.models import Category


class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = [
            'id',
            'name',
        ]